feat: 무한 슬라이드 버튼 이벤트 적용 완료
