+++ src/scripts/utils/carousel/set_transiton.js
@@ -1,0 +1,3 @@
+export function setTransition(on, el) {
+  el.style.transition = on ? 'transform 0.4s ease-in-out' : 'none';
+}
